+++ lib_mic_array/api/mic_array.h
@@ -130,6 +130,8 @@ typedef unsigned mic_array_internal_audio_channels;
  *  \param c_from_pdm_interface      The channel where the decimated PDM from pdm_rx task will be inputted.
  *  \param c_frame_output            The channel used to transfer data and control information between
  *                                   the client of this task and this task.
+ *  \param channels                  A pointer to an array of mic_array_internal_audio_channels. This can be set to
+ *                                   MIC_ARRAY_NO_INTERNAL_CHANS if none are requires.
  */
 void mic_array_decimate_to_pcm_4ch(
         streaming chanend c_from_pdm_interface,
@@ -141,13 +143,20 @@ void mic_array_decimate_to_pcm_4ch(
  *  array. The sample rate of the producer and sample rate of the output of the microphone array must
  *  match.
  *
- *  \param c_from_pdm_interface      The channel where the decimated PDM from pdm_rx task will be inputted.
- *  \param c_frame_output            The channel used to transfer data and control information between
- *                                   the client of this task and this task.
+ *  \param internal_channels         An array of mic_array_internal_audio_channels.
+ *                                   MIC_ARRAY_NO_INTERNAL_CHANS if none are requires.
+ *  \param ch0                       The channel used to send internal audio to mic_array
+ *                                   channel 0.
+ *  \param ch1                       The channel used to send internal audio to mic_array
+ *                                   channel 1.
+ *  \param ch2                       The channel used to send internal audio to mic_array
+ *                                   channel 2.
+ *  \param ch3                       The channel used to send internal audio to mic_array
+ *                                   channel 3.
  */
 void mic_array_init_far_end_channels(mic_array_internal_audio_channels internal_channels[4],
-        streaming chanend ?a, streaming chanend ?b,
-        streaming chanend ?c, streaming chanend ?d);
+        streaming chanend ?ch0, streaming chanend ?ch1,
+        streaming chanend ?ch2, streaming chanend ?ch3);
 
 /** This sends an audio sample to a decimator.
  *