+++ lib_mic_array/api/mic_array.h
@@ -11,6 +11,9 @@
     #define MIC_ARRAY_HIRES_MAX_DELAY 256
 #endif
 
+#define MIC_ARRAY_NO_INTERNAL_CHANS (0)
+
+
 /** PDM Microphone Interface component.
  *
  *  This task handles the interface to up to 8 PDM microphones whilst also decimating
@@ -113,6 +116,8 @@ typedef struct {
 
 } mic_array_decimator_config_t;
 
+typedef unsigned mic_array_internal_audio_channels;
+
 /** Four Channel Decimation component.
  *
  *  This task decimates the four channel input down to the desired output sample rate.
@@ -125,11 +130,48 @@ typedef struct {
  *  \param c_from_pdm_interface      The channel where the decimated PDM from pdm_rx task will be inputted.
  *  \param c_frame_output            The channel used to transfer data and control information between
  *                                   the client of this task and this task.
+ *  \param channels                  A pointer to an array of mic_array_internal_audio_channels. This can be set to
+ *                                   MIC_ARRAY_NO_INTERNAL_CHANS if none are requires.
  */
 void mic_array_decimate_to_pcm_4ch(
         streaming chanend c_from_pdm_interface,
-        streaming chanend c_frame_output);
+        streaming chanend c_frame_output, mic_array_internal_audio_channels * channels);
 
+/** Far end channel connector.
+ *
+ *  This function allowed a connection to be established between a signal producer and the microphone
+ *  array. The sample rate of the producer and sample rate of the output of the microphone array must
+ *  match.
+ *
+ *  \param internal_channels         An array of mic_array_internal_audio_channels.
+ *                                   MIC_ARRAY_NO_INTERNAL_CHANS if none are requires.
+ *  \param ch0                       The channel used to send internal audio to mic_array
+ *                                   channel 0.
+ *  \param ch1                       The channel used to send internal audio to mic_array
+ *                                   channel 1.
+ *  \param ch2                       The channel used to send internal audio to mic_array
+ *                                   channel 2.
+ *  \param ch3                       The channel used to send internal audio to mic_array
+ *                                   channel 3.
+ */
+void mic_array_init_far_end_channels(mic_array_internal_audio_channels internal_channels[4],
+        streaming chanend ?ch0, streaming chanend ?ch1,
+        streaming chanend ?ch2, streaming chanend ?ch3);
+
+/** This sends an audio sample to a decimator.
+ *
+ *  This function call sets up the four channel decimators. After this has been called there
+ *  will be a real time requirement on this task, i.e. this task must call
+ *  mic_array_get_next_time_domain_frame() at the output sample rate multiplied by the frame size.
+ *
+ *  \param c_to_decimator    The channel used to transfer audio sample beterrn the application and
+ *                           the decimators.
+ *  \param sample            The audio sample to be transfered.
+ *  \returns                 0 for success and 1 for failure. Failure may occour when the decimators
+ *                           are not yet running.
+ *
+ */
+int mic_array_send_sample( streaming chanend c_to_decimator, int sample);
 
 /** Four Channel Decimation initializer for raw audio frames.
  *
@@ -157,6 +199,13 @@ void mic_array_init_time_domain_frame(
  *  This function handles the frame exchange between the mic_array_decimate_to_pcm_4ch() tasks and the
  *  application. It returns a pointer to the most recently written frame. After this point the oldest
  *  frame is assumed out of scope to the application.
+ *  Current behaviour: this function must be called before the decimators have produced the full frame
+ *  and are ready for exachange. If this function is not called in time then the PDM data will be dropped
+ *  until the frame is accepted from the decimators. Failure to meet timing can be detected by enabling
+ *  DEBUG_MIC_ARRAY in your makefile with -DDEBUG_MIC_ARRAY.
+ *  Future behaviour: if this function is not called before the frame has been produced by the decimators
+ *  then the frame is dropped. This can be detected with the frame_counter attached to the metadata of the
+ *  frame. Sequential frames will have frame_counters that increament by one.
  *
  *  \param c_from_decimators The channels used to transfer pointers between the application and
  *                           the mic_array_decimate_to_pcm_4ch() tasks.
@@ -196,6 +245,13 @@ void mic_array_init_frequency_domain_frame(streaming chanend c_from_decimators[]
  *  This function handles the frame exchange between the mic_array_decimate_to_pcm_4ch() tasks and the
  *  application. It returns a pointer to the most recently written frame. After this point the oldest
  *  frame is assumed out of scope to the application.
+ *  Current behaviour: this function must be called before the decimators have produced the full frame
+ *  and are ready for exachange. If this function is not called in time then the PDM data will be dropped
+ *  until the frame is accepted from the decimators. Failure to meet timing can be detected by enabling
+ *  DEBUG_MIC_ARRAY in your makefile with -DDEBUG_MIC_ARRAY.
+ *  Future behaviour: if this function is not called before the frame has been produced by the decimators
+ *  then the frame is dropped. This can be detected with the frame_counter attached to the metadata of the
+ *  frame. Sequential frames will have frame_counters that increament by one.
  *
  *  \param c_from_decimators   The channels used to transfer pointers between the application and
  *                            the mic_array_decimate_to_pcm_4ch() tasks.