+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ lib_mic_array change log
 -----
 
   * Added ability to route internal channels of the output rate of the mic_array
-    to the mic_array so that they can benefit from the post processing of the 
+    to the mic_array so that they can benefit from the post processing of the
     mic_array.
 
 2.0.1